chore: split mural journal sync service
