+++ docs/agent-audit/reasoning/2026/06/06/reduce-repository-debt.json
@@ -68,6 +68,16 @@
     "infra/cloudflare/src/service/projects/csv.js",
     "infra/cloudflare/src/service/projects/github-csv.js",
     "infra/cloudflare/src/service/projects/normalisation.js",
+    "infra/cloudflare/src/service/mural-journal-sync.js",
+    "infra/cloudflare/src/service/mural-journal-sync/auth.js",
+    "infra/cloudflare/src/service/mural-journal-sync/constants.js",
+    "infra/cloudflare/src/service/mural-journal-sync/context.js",
+    "infra/cloudflare/src/service/mural-journal-sync/mural-api.js",
+    "infra/cloudflare/src/service/mural-journal-sync/request.js",
+    "infra/cloudflare/src/service/mural-journal-sync/sticky-payloads.js",
+    "infra/cloudflare/src/service/mural-journal-sync/sync-entry.js",
+    "infra/cloudflare/src/service/mural-journal-sync/text.js",
+    "infra/cloudflare/src/service/mural-journal-sync/widgets.js",
     "lychee.toml",
     "public/components/guides/api.js",
     "public/components/guides/front-matter.js",
@@ -76,6 +86,7 @@
     "public/partials/project-tabs.html",
     "tests/ai-rewrite-split-route-state.test.js",
     "tests/govuk-pages-render-workflow-state.test.js",
+    "tests/mural-journal-sync-route-state.test.js",
     "tests/mural-service-split-route-state.test.js",
     "tests/projects-service-split-route-state.test.js",
     "tests/study-guides-route-state.test.js",
@@ -91,7 +102,8 @@
     "Follow-up Lychee remediation corrected stale documentation links exposed by the narrower Git ignore rules, fixed the project notes tab route, and added narrow Lychee exclusions for placeholder-heavy Mural example transcripts and the Mustache HTML head partial.",
     "Follow-up Mural service split kept MuralServicePart as the public route surface while moving board registry resolution, direct journal sticky writing, token refresh, viewer URL probing, and workspace/room setup into responsibility-specific modules.",
     "Follow-up AI rewrite split kept ai-rewrite.js as the public route and Worker entry surface while moving defaults, HTTP helpers, text sanitisation, guardrails, prompt content, and testing helpers into responsibility-specific modules under infra/cloudflare/src/core/ai-rewrite/.",
-    "Follow-up projects service split kept projects.js as the public route operation layer while moving CSV parsing, GitHub CSV fallback, field normalisation and mapping, auth visibility checks, and Airtable access helpers into responsibility-specific modules under infra/cloudflare/src/service/projects/."
+    "Follow-up projects service split kept projects.js as the public route operation layer while moving CSV parsing, GitHub CSV fallback, field normalisation and mapping, auth visibility checks, and Airtable access helpers into responsibility-specific modules under infra/cloudflare/src/service/projects/.",
+    "Follow-up Mural journal sync split kept mural-journal-sync.js as the public route handler while moving constants, text and tag parsing, widget and template matching, sticky payload shaping, Mural API calls, auth and token handling, request parsing, context building, and per-entry sync into responsibility-specific modules under infra/cloudflare/src/service/mural-journal-sync/."
   ],
   "validation": [
     {
@@ -205,6 +217,34 @@
     {
       "command": "npm run trace:coverage",
       "status": "passed_after_projects_service_split"
+    },
+    {
+      "command": "node --check infra/cloudflare/src/service/mural-journal-sync.js && for f in infra/cloudflare/src/service/mural-journal-sync/*.js; do node --check \"$f\"; done",
+      "status": "passed_after_mural_journal_sync_split"
+    },
+    {
+      "command": "node --test tests/mural-journal-sync-route-state.test.js",
+      "status": "passed_after_mural_journal_sync_split"
+    },
+    {
+      "command": "dynamic import of infra/cloudflare/src/service/mural-journal-sync.js with public export assertion",
+      "status": "passed_after_mural_journal_sync_split"
+    },
+    {
+      "command": "npm test",
+      "status": "passed_181_tests_after_mural_journal_sync_split"
+    },
+    {
+      "command": "npm run lint",
+      "status": "passed_with_existing_warnings_after_mural_journal_sync_split"
+    },
+    {
+      "command": "npm run format:check",
+      "status": "passed_after_generated_css_regeneration_after_mural_journal_sync_split"
+    },
+    {
+      "command": "npm run trace:coverage",
+      "status": "passed_after_mural_journal_sync_split"
     }
   ],
   "residualRisks": [
