+++ docs/agent-audit/reasoning/2026/06/06/reduce-repository-debt.md
@@ -81,6 +81,16 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - infra/cloudflare/src/service/projects/csv.js
 - infra/cloudflare/src/service/projects/github-csv.js
 - infra/cloudflare/src/service/projects/normalisation.js
+- infra/cloudflare/src/service/mural-journal-sync.js
+- infra/cloudflare/src/service/mural-journal-sync/auth.js
+- infra/cloudflare/src/service/mural-journal-sync/constants.js
+- infra/cloudflare/src/service/mural-journal-sync/context.js
+- infra/cloudflare/src/service/mural-journal-sync/mural-api.js
+- infra/cloudflare/src/service/mural-journal-sync/request.js
+- infra/cloudflare/src/service/mural-journal-sync/sticky-payloads.js
+- infra/cloudflare/src/service/mural-journal-sync/sync-entry.js
+- infra/cloudflare/src/service/mural-journal-sync/text.js
+- infra/cloudflare/src/service/mural-journal-sync/widgets.js
 - public/components/guides/api.js
 - public/components/guides/front-matter.js
 - public/components/guides/guides-page.js
@@ -89,6 +99,7 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - tests/govuk-pages-render-workflow-state.test.js
 - tests/ai-rewrite-split-route-state.test.js
 - tests/mural-service-split-route-state.test.js
+- tests/mural-journal-sync-route-state.test.js
 - tests/projects-service-split-route-state.test.js
 - tests/study-guides-route-state.test.js
 - docs/agent-audit/reasoning/2026/06/06/reduce-repository-debt.md
@@ -105,6 +116,7 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - Follow-up Mural service split kept `MuralServicePart` as the public route surface while moving board registry resolution, direct journal sticky writing, token refresh, viewer URL probing, and workspace/room setup into responsibility-specific modules.
 - Follow-up AI rewrite split kept `ai-rewrite.js` as the public route and Worker entry surface while moving defaults, HTTP helpers, text sanitisation, guardrails, prompt content, and testing helpers into responsibility-specific modules under `infra/cloudflare/src/core/ai-rewrite/`.
 - Follow-up projects service split kept `projects.js` as the public route operation layer while moving CSV parsing, GitHub CSV fallback, field normalisation/mapping, auth visibility checks, and Airtable access helpers into responsibility-specific modules under `infra/cloudflare/src/service/projects/`.
+- Follow-up Mural journal sync split kept `mural-journal-sync.js` as the public route handler while moving constants, text/tag parsing, widget/template matching, sticky payload shaping, Mural API calls, auth/token handling, request parsing, context building, and per-entry sync into responsibility-specific modules under `infra/cloudflare/src/service/mural-journal-sync/`.
 
 ## Validation attempted
 
@@ -136,6 +148,13 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - Follow-up projects service split: `npm run lint` — passed with existing warnings only.
 - Follow-up projects service split: `npm run format:check` — passed.
 - Follow-up projects service split: `npm run trace:coverage` — passed.
+- Follow-up Mural journal sync split: `node --check infra/cloudflare/src/service/mural-journal-sync.js && for f in infra/cloudflare/src/service/mural-journal-sync/*.js; do node --check "$f"; done` — passed.
+- Follow-up Mural journal sync split: `node --test tests/mural-journal-sync-route-state.test.js` — passed.
+- Follow-up Mural journal sync split: dynamic import of `infra/cloudflare/src/service/mural-journal-sync.js` confirmed the public export remains available — passed.
+- Follow-up Mural journal sync split: `npm test` — passed, 181 tests.
+- Follow-up Mural journal sync split: `npm run lint` — passed with existing warnings only.
+- Follow-up Mural journal sync split: `npm run format:check` — passed after generated CSS was regenerated by lint.
+- Follow-up Mural journal sync split: `npm run trace:coverage` — passed.
 
 ## Residual risk
