+++ infra/cloudflare/src/service/mural-journal-sync/constants.js
@@ -1,0 +1,6 @@
+export const CATEGORY_KEYS = ["perceptions", "procedures", "decisions", "introspections"];
+export const DEFAULT_WIDTH = 260;
+export const DEFAULT_HEIGHT = 160;
+export const GRID_GAP = 32;
+export const PURPOSE_REFLEXIVE = "reflexive_journal";
+export const TEMPLATE_PLACEHOLDER_RE = /This is a .*sticky note that will contain|Further .*entries will appear|single .*from the Research Operations app|Research Operations app/i;
