+++ tests/mural-journal-sync-route-state.test.js
@@ -2,6 +2,14 @@ import assert from "node:assert/strict";
 import fs from "node:fs";
 
 const serviceSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync.js", "utf8");
+const baseConstantsSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync/constants.js", "utf8");
+const baseContextSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync/context.js", "utf8");
+const baseMuralApiSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync/mural-api.js", "utf8");
+const basePayloadsSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync/sticky-payloads.js", "utf8");
+const baseRequestSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync/request.js", "utf8");
+const baseSyncEntrySource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync/sync-entry.js", "utf8");
+const baseTextSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync/text.js", "utf8");
+const baseWidgetsSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync/widgets.js", "utf8");
 const layoutSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync-layout.js", "utf8");
 const safeTagsSource = fs.readFileSync("infra/cloudflare/src/service/mural-journal-sync-safe-tags.js", "utf8");
 const indexSource = fs.readFileSync("infra/cloudflare/src/service/index.js", "utf8");
@@ -36,29 +44,42 @@ function includesElementAttributes(source, elementName, expectedAttributes, labe
 includes(serviceSource, "export async function muralJournalSync", "service");
 includes(serviceSource, "mode === \"status\"", "service");
 includes(serviceSource, "mode === \"hydrate\"", "service");
-includes(serviceSource, "journal-entry:", "service");
-includes(serviceSource, "widgetHasEntryTag", "service");
-includes(serviceSource, "canonicalExistingWidget", "service");
-includes(serviceSource, "categoryTemplateWidget", "service");
-includes(serviceSource, "candidateTemplateWidgets", "service");
-includes(serviceSource, "latestCanonicalCategoryWidget", "service");
-includes(serviceSource, "widgetMatchesTemplateGeometry", "service");
-includes(serviceSource, "tagsForEntry", "service");
-includes(serviceSource, "stickyStyleFromTemplate", "service");
-includes(serviceSource, "stickyPayloadFromTemplate", "service");
-includes(serviceSource, "createStickyFromTemplate", "service");
-includes(serviceSource, "widgets/sticky-note", "service");
 includes(serviceSource, "updated-template-widget", "service");
 includes(serviceSource, "created-template-sticky", "service");
 includes(serviceSource, "createdOrUpdated", "service");
 includes(serviceSource, "function hydrateReason", "service");
-includes(serviceSource, "template-sticky-minimal-array", "service");
-includes(serviceSource, "widgetMetadataText", "service");
-includes(serviceSource, "title: syncTitle", "service");
-includes(serviceSource, "#FFFFFFFF", "service");
-includes(serviceSource, "tags: patch.tags", "service");
 includes(serviceSource, "errors: err?.errors", "service");
 includes(serviceSource, "reason: hydrateReason", "service");
+includes(serviceSource, "./mural-journal-sync/context.js", "service");
+includes(serviceSource, "./mural-journal-sync/request.js", "service");
+includes(serviceSource, "./mural-journal-sync/sync-entry.js", "service");
+includes(serviceSource, "./mural-journal-sync/text.js", "service");
+
+includes(baseConstantsSource, "export const CATEGORY_KEYS", "base constants");
+includes(baseConstantsSource, "PURPOSE_REFLEXIVE", "base constants");
+includes(baseTextSource, "export function normalizeCategoryKey", "base text helpers");
+includes(baseTextSource, "export function normalizeTags", "base text helpers");
+includes(baseWidgetsSource, "journal-entry:", "base widget helpers");
+includes(baseWidgetsSource, "export function widgetHasEntryTag", "base widget helpers");
+includes(baseWidgetsSource, "export function categoryTemplateWidget", "base widget helpers");
+includes(baseWidgetsSource, "export function candidateTemplateWidgets", "base widget helpers");
+includes(baseWidgetsSource, "export function latestCanonicalCategoryWidget", "base widget helpers");
+includes(baseWidgetsSource, "export function widgetMatchesTemplateGeometry", "base widget helpers");
+includes(baseWidgetsSource, "export function widgetMetadataText", "base widget helpers");
+includes(basePayloadsSource, "export function tagsForEntry", "base sticky payload helpers");
+includes(basePayloadsSource, "export function stickyStyleFromTemplate", "base sticky payload helpers");
+includes(basePayloadsSource, "export function stickyPayloadFromTemplate", "base sticky payload helpers");
+includes(basePayloadsSource, "title: syncTitle", "base sticky payload helpers");
+includes(basePayloadsSource, "#FFFFFFFF", "base sticky payload helpers");
+includes(baseMuralApiSource, "export async function createStickyFromTemplate", "base Mural API helpers");
+includes(baseMuralApiSource, "widgets/sticky-note", "base Mural API helpers");
+includes(baseMuralApiSource, "template-sticky-minimal-array", "base Mural API helpers");
+includes(baseMuralApiSource, "tags: patch.tags", "base Mural API helpers");
+includes(baseRequestSource, "export function parseEntryPayload", "base request helpers");
+includes(baseContextSource, "export async function buildContext", "base context helpers");
+includes(baseContextSource, "export function statusFromEntriesAndWidgets", "base context helpers");
+includes(baseSyncEntrySource, "export function canonicalExistingWidget", "base entry sync helpers");
+includes(baseSyncEntrySource, "export async function syncOneEntry", "base entry sync helpers");
 
 includes(layoutSource, "export async function muralJournalSync", "layout service");
 includes(layoutSource, "function categoryHeaderWidget", "layout service");
